+++ src/eca/features/skills.clj
@@ -28,7 +28,7 @@
 
 (defn global-skills-dir []
   (let [xdg-config-home (or (config/get-env "XDG_CONFIG_HOME")
-                            (io/file (config/get-property "user.home") ".config"))]
+                            (io/file (config/user-home) ".config"))]
     (io/file xdg-config-home "eca" "skills")))
 
 (defn ^:private skill-file? [file]